requirements.txt
pyyaml
pandas
numpy
matplotlib
seaborn
pytest
pytest-cov
python-dateutil
scipy
